requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-docx==1.1.0
pymupdf==1.24.0
PyMuPDFb==1.24.0
textract==1.6.5
requests==2.31.0
supabase==2.3.4
python-multipart==0.0.6
pydantic==2.5.0